Reorganize customer grid/card files into own components, clean up i18n
