+++ eisbuk-admin/src/components/atoms/CustomerCard/CustomerCard.stories.tsx
@@ -1,15 +1,15 @@
 import React from "react";
 import { ComponentMeta } from "@storybook/react";
 
-import CustomerDialog from "./CustomerDialog";
+import CustomerCard from "./CustomerCard";
 
 import { walt } from "@/__testData__/customers";
 
 export default {
   title: "Customer Dialog",
-  component: CustomerDialog,
-} as ComponentMeta<typeof CustomerDialog>;
+  component: CustomerCard,
+} as ComponentMeta<typeof CustomerCard>;
 
 export const Default = (): JSX.Element => (
-  <CustomerDialog customer={walt} onClose={(): void => {}} />
+  <CustomerCard customer={walt} onClose={(): void => {}} />
 );
