+++ eisbuk-admin/src/components/atoms/CustomerCard/index.ts
@@ -1,0 +1,3 @@
+import CustomerCard from "./CustomerCard";
+
+export default CustomerCard;
