+++ eisbuk-admin/src/components/atoms/CustomerCard/types.ts
@@ -1,0 +1,7 @@
+import { Customer } from "eisbuk-shared";
+
+export interface ActionButtonProps {
+  customer: Customer;
+  onClose: () => void;
+  className?: string;
+}
