+++ eisbuk-admin/src/components/atoms/CustomerGrid/index.ts
@@ -1,0 +1,3 @@
+import CustomerGrid from "./CustomerGrid";
+
+export default CustomerGrid;
